refactor(order): separate controller service and repository layers
