+++ apps/api/src/strategies/cashPaymentStrategy.ts
@@ -1,0 +1,14 @@
+import { PaymentStrategy } from "./paymentStrategy";
+
+export class CashPaymentStrategy implements PaymentStrategy {
+  async pay(amount: number) {
+    if (amount <= 0) {
+      throw new Error("Invalid payment amount");
+    }
+
+    return {
+      method: "CASH" as const,
+      status: "PAID" as const,
+    };
+  }
+}
