+++ apps/api/src/strategies/qrisPaymentStrategy.ts
@@ -1,0 +1,14 @@
+import { PaymentStrategy } from "./paymentStrategy";
+
+export class QrisPaymentStrategy implements PaymentStrategy {
+  async pay(amount: number) {
+    if (amount <= 0) {
+      throw new Error("Invalid payment amount");
+    }
+
+    return {
+      method: "QRIS" as const,
+      status: "PAID" as const,
+    };
+  }
+}
